requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.0                # via pip-chill, pip-tools
dj-database-url==0.5.0
django-filter==2.2.0
django==2.2.6
djangorestframework==3.10.3
gunicorn==19.9.0
pip-chill==0.1.8
pip-tools==4.1.0
psycopg2==2.8.3
pytz==2019.3              # via django
six==1.12.0               # via pip-tools
sqlparse==0.3.0           # via django
whitenoise==4.1.4
